Тестовое Задание/requirements.txt
aiogram>=3.4.0
psycopg2-binary>=2.9.10
SQLAlchemy>=2.0.25
alembic==1.13.1
python-dotenv==1.0.0
loguru==0.7.2
pydantic>=2.12.0
pandas>=2.2.0
openpyxl==3.1.2
aiohttp==3.9.1
mypy==1.8.0
